chromeos-bootimage: Add an option to include U-Boot
U-Boot is a boot loader widely used on ARM devices. It also has support
for some x86 boards. It is useful to be able to include this in a firmware
image, for hardware testing and developer-mode flexibility. Add a USE flag
to install it.

BUG=chromium:837018
TEST=USE=u_boot emerge-grunt chromeos-bootimage
See that it includes U-Boot:
$ cbfstool  /build/grunt/firmware/image-grunt.serial.bin print -r RW_LEGACY
Name                           Offset     Type           Size   Comp
u-boot                         0x0        simple elf     221535 none
(empty)                        0x361c0    null          7642584 none

Change-Id: I095fe0a330b6ce9424e0f08110c9564072456771
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1010840
Reviewed-by: Raul E Rangel <rrangel@chromium.org>