pppd: Patches to allow starting as non-root user.
This CL is in preparation to run the shill process tree as a non-root
user. It doesn't change any current functionality, but will let us run
ppp as user "shill" in the future.

In order to run pppd as a non-root user, we need to do two things:
 1) Omit checks in startup code that asserts that pppd is running
    with root privs.
 2) Allow to configure pppd at compile time to use a different runtime
    data directory than the default, which is /var/run, since pppd will
    soon run in a minijail under 'shill' user and we do not want to give
    'shill' user access to /var/run.

Patch #2 isn't in appropriate shape to upstream, so more work would need
to be done there if we want to do that. That being said, the last
version of ppp to be released was in 2014.

BUG=chromium:649417
TEST=tested as part of larger sandbox shill debug CL

Change-Id: Ifa9f2f785e74b9de8f3b9f4bc4f68c53d8b8e0ed
Reviewed-on: https://chromium-review.googlesource.com/1022398
Commit-Ready: Micah Morton <mortonm@chromium.org>
Tested-by: Micah Morton <mortonm@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Kevin Cernekee <cernekee@chromium.org>