chromeos-init: don't force debugd group on non-debugd builds
Embedded devices may not run with debugd, so we shouldn't assume so in
the pstore init script. This causes the pstore "service" to fail to come
up, and we don't end up with a proper /dev/pstore/ symlink.

Seen on wifi-testbed builds.

Works in conjunction with CL:1027077

BUG=none
TEST=see pstore come up on wifi-testbed and base profiles

Change-Id: Iae027cd9c1f530f274ef3cce5135ec16670af706
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1027004
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>