linux-headers: Add linux/gpio.h
Copy the linux kernel's gpio.h header file from v4.14 and export it into
the UAPI (which happened in v4.6) as one squashed patch so we don't have
to stack the 6 incremental updates to this header as more patch files.
This is used to compile libgpiod and use the userspace gpiodev
interface.

BUG=chromium:839019
TEST=emerge linux-headers
Change-Id: Ie5baf010ec522ff93c519ef0582ba0d2f6bdcade
Reviewed-on: https://chromium-review.googlesource.com/1040667
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Alexandru M Stan <amstan@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>