wpa_supplicant: fix syntax in DBUS perms file
This CL is in preparation to run the shill process tree as a non-root
user.

This change hasn't had any effect since we are not yet running shill as
'shill' user, but these 2 lines were using the wrong syntax which meant
'shill' user was not actually getting the permissions it appeared to.

BUG=chromium:649417
TEST=tested as part of larger shill sandboxing debug CL

Change-Id: Icef756970dbb88e240e3a8c4dc0973514a80e7ee
Reviewed-on: https://chromium-review.googlesource.com/1038732
Commit-Ready: Micah Morton <mortonm@chromium.org>
Tested-by: Micah Morton <mortonm@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>