Add -fsanitize-blacklist when CFI is enabled.
In order to work properly CFI needs to use a particular
blacklist.  LLVM used to find the blacklist implicitly, but
the location changed recently and goma builds are looking for
the blacklist in the wrong place.  With this CL, we explicitly
tell LLVM where the blacklist file is.

BUG=chromium:838428
TEST=Launched two tryjobs on terra-release-tryjob, enabling CFI
on both, but only passing this CL to one of them.  The job with
this CL generated a working chromiumos image; the one without
this CL generated a broken chromiumos image.

Change-Id: I5abaae3cc0c6d32dfa64f517a01fd8ed5f3ed28d
Reviewed-on: https://chromium-review.googlesource.com/1036923
Commit-Ready: Caroline Tice <cmtice@chromium.org>
Tested-by: Caroline Tice <cmtice@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>