chromeos-chrome: remove -fdebug-info-for-profiling for ARM
We added the -fdebug-info-for-profiling flag because we want to get AFDO
profile from perf. Currently we do not use ARM machine to collect
profiles. so this flag is useless for arm. In addition, it caused an
LLVM assertion when thinlto and debug fission is use. While working on
LLVM to have a fix there, we want to remove this flag so that we can
test thinlto for ARM.

BUG=chromium:834839
TEST=the llvm assertion is gone.

Change-Id: Ie768a79e1cffb8eeb3d85ae1bdfcedd5e6f3c27c
Reviewed-on: https://chromium-review.googlesource.com/1041652
Commit-Ready: Yunlian Jiang <yunlian@chromium.org>
Tested-by: Yunlian Jiang <yunlian@chromium.org>
Reviewed-by: Luis Lozano <llozano@chromium.org>