toolchain-funcs.eclass: fix tc-ld-disable-gold when using clang
tc-ld-disable-gold checks gcc version to see if we have gcc-4.8+
The version check fails if clang is set as the compiler.
  $ clang -E -P - <<<"__GNUC__ __GNUC_MINOR__ __GNUC_PATCHLEVEL__"
  4 2 1
i.e. clang returns a gcc version of 4.2.1
This results in incorrectly adding -B ... to LDFLAGS, when clang
supports "-fuse-ld" just fine.

Support for "-fuse-ld" first appeared in clang-3.5, so check clang
version and use the flag if supported.

BUG=chromium:839665
TEST=emerge-guado_moblab sys-process/criu

Change-Id: I6e207bd8875effe308958e165d78cdd59adff3fb
Reviewed-on: https://chromium-review.googlesource.com/1045553
Commit-Ready: Rahul Chaudhry <rahulchaudhry@chromium.org>
Tested-by: Rahul Chaudhry <rahulchaudhry@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>