iwlwifi_rescan: Enable L1 and disable L1 substates after rescan
I noticed that L1 is disabled by the kernel when device is removed
but not enabled again on a rescan. Lets fix that manually because
the PCI maintainer would like to keep such policy in user space anyway)
(Also slightly related: https://patchwork.kernel.org/patch/10387751/)

Also, the L1 substates are disabled in BIOS, lets disable it again after
a rediscovery because there is no BIOS at this time to disable (this is
only a security measure as they are not enabled by default).

BUG=b:79427560
TEST=verified that after rescan the bits are as expected.
Change-Id: I28f35934f0e3e695480d0e8d48fafb78cdeeba47
Reviewed-on: https://chromium-review.googlesource.com/1050990
Commit-Ready: Rajat Jain <rajatja@chromium.org>
Tested-by: Rajat Jain <rajatja@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>