eclass: Enable building the kernel with clang for arm64
Add arm64 conditions to cros-kernel2.eclass in order to
the kernel image can be built properly and contains FDT as well.

Enable the build for arm64 with clang in toolchain-funcs.eclass.

Also did some style fixes in the get_build_arch function
to be consistent with the proposal.

BUG=None
TEST=emerge-arm64-generic chromeos-kernel-4_4

Change-Id: I943f6245baff4871723ca2aff94cbfd6e490e7b7
Reviewed-on: https://chromium-review.googlesource.com/1052691
Commit-Ready: Adam Kallai <kadam@inf.u-szeged.hu>
Tested-by: Adam Kallai <kadam@inf.u-szeged.hu>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>