conntrack-tools: fix init tracking
minijail0 & conntrackd both stay in the foreground, so "expect daemon"
doesn't work as it expects a third process to fork and the others to
exit.  Add -i to minijail0 so that the first minijail0 exits, and change
to "expect fork" so the conntrackd process is tracked by upstart.

BUG=chromium:684014
TEST=start/stop works in a loop

Change-Id: Ic3ed00fcb6ff447301a8f896da1a3545623c7ed9
Reviewed-on: https://chromium-review.googlesource.com/1053432
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Kevin Cernekee <cernekee@chromium.org>