dev-lang/rust: upgrade to rust 1.26
This version saw a lot of changes thanks to the target triple no longer
being a simple string. There were 3 patches which had to be adjusted
accordingly.

Additionally, there was a bug fix with the +fp feature on aarch64 that
didn't make it into the 1.26 release, which is cherry-picked here.

This change also removes the armv7a_cros_linux_gnueabi.rs file, which
has no use as far as I can tell.

BUG=None
TEST=sudo emerge rust

Change-Id: I2d7311d8a33f4cc28071d4c058ba38a60317868a
Reviewed-on: https://chromium-review.googlesource.com/1056406
Commit-Ready: Jason Clinton <jclinton@chromium.org>
Tested-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Jason Clinton <jclinton@chromium.org>