Prepare for update fontconfig to 2.13.0
Add a copy of unchanged upstream ebuild so that our local
changes on top of it can be recorded.

It's masked so that we'll keep using fontconfig-2.11 before it's
unmasked along with local changes.

There are two local changes, nonetheless

1. eapi7-ver eclass is not present so that it's dropped from inherit
line. The check for the 3rd component of the version number (
[[ $(ver_cut 3) -ge 90 ]] ) before KEYWORDS is also dropped because it
uses ver_cut from eapi7-ver.

2. KEYWORDS is set to '~*' to pass the presubmit check.

BUG=chromium:747459
TEST=None (fc 2.13 is masked)

Change-Id: I4d5f8ad46e616fa49ed0cf7a2b11f46e8b6e3863
Reviewed-on: https://chromium-review.googlesource.com/1065214
Commit-Ready: Jungshik Shin <jshin@chromium.org>
Tested-by: Jungshik Shin <jshin@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>