Upgrade grub from 2.00 to 2.02
The 2.00 version of the ebuild used a git repo that contained a copy
of the contents of the 2.00 tarball, plus autogenerated build
files. There were only a couple of CrOS-specific patches in that repo.

The new 2.02 ebuild no longer uses that git repo, so it isn't
cros-workon-able now. Instead it uses the 2.02 release tarball with a
couple of CrOS patches applied. It also calls `autogen.sh` now so that
the applied patches don't need to include autogenerated build files.

BUG=None
TEST=sudo emerge grub && build_image, verify that both UEFI and legacy BIOS booting works

Change-Id: I2fe837f4bc3f6ec4cab75b2e92628ef671eba43f
Reviewed-on: https://chromium-review.googlesource.com/1066750
Commit-Ready: Nicholas Bishop <nbishop@neverware.com>
Tested-by: Nicholas Bishop <nbishop@neverware.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>