libchrome: Add FuzzedDataProvider.
Chrome OS is ramping up the use of fuzzing for increased stability and
security. We published a fuzzing guide at
https://chromium.googlesource.com/chromiumos/docs/+/master/fuzzing.md
and now it's time to start writing fuzzers.

FuzzedDataProvider simplifies writing fuzzers that test more than just
functions that take a |char*|. It makes it a lot easier to test APIs by
allowing the fuzzer to use the data provided by the fuzzing scaffolding
to make higher level decisions.

Add FuzzedDataProvider to libchrome to allow Chrome OS fuzzers to use
it.

This is a cherry-pick of CL:1069002

BUG=chromium:844592
TEST=Compiles, fuzzer can use FuzzedDataProvider.

Change-Id: Ib7686122ee42add55691bf69574a2cbf4a5812f0
Reviewed-on: https://chromium-review.googlesource.com/1070435
Commit-Ready: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>