dev-lang/rust: check toolchains only when building in pkg_setup
Instead of failing rather late in the emerge process of the rust
package, this change adds a check very early with a precise error
message if a required toolchain is missing. This check is skipped if the
package is not being built locally because it is a binpkg.

TEST=emerge rust
BUG=chromium:842557

Change-Id: I203bda9044946b7691bcaf80883db13415802563
Reviewed-on: https://chromium-review.googlesource.com/1067505
Commit-Ready: Zach Reizner <zachr@chromium.org>
Tested-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>