Replace chromeos-base/root-certificates with app-misc/ca-certificates
The former existed at a time when Debian and Gentoo did not have a
policy of keeping their certificates in sync with the Mozilla NSS
(Network Security Services) module - the list of CA certificates that
Chrome trusts. That is not the case any more and we are comfortable with
working with Gentoo to keep app-misc/ca-certificates up to date. Hence
this CL.

This drops chromeos-base/root-certificates, replaces it with
app-misc/ca-certificates in implicit-system, removes
app-misc/ca-certificates from the provided list of the chromeos target,
and from the SDK's RDEPEND - it will still by pulled in if there are
packages that depend on it.

BUG=b:78572490
BUG=chromium:320802
TEST=`emerge-lakitu app-misc/ca-certificates` doesn't say it's provided
    any more, emerging implicit-system also pulls it in.
CQ-DEPEND=CL:1070479,CL:1083226

Change-Id: I3f930cd23f8e5cbef0e916094689d99c1d61fef5
Reviewed-on: https://chromium-review.googlesource.com/1070605
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Daniel Wang <wonderfly@google.com>
Reviewed-by: Ryan Sleevi <rsleevi@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>