dev-go/cdp: Rename 1.18.1 to 0.18.1.
It looks like there was a mistake upstream that resulted in
a v1.18.1 tag being created in place of v0.18.1.
a66b67bfc0d3 updated our ebuild to the "new" version. The
v1.18.1 tag no longer exists in the upstream repo (or in our
mirror of it), so switch to the v0.18.1 tag.

BUG=chromium:846369
TEST=none

Change-Id: Ibe11c1565f2da8a20dd83c946755a5a36bd008f0
Reviewed-on: https://chromium-review.googlesource.com/1072897
Tested-by: Dan Erat <derat@chromium.org>
Trybot-Ready: Dan Erat <derat@chromium.org>
Reviewed-by: Shen-En Shih <petershih@chromium.org>
Commit-Queue: Shen-En Shih <petershih@chromium.org>