Remove fps_meter ebuild files
This patch is a part of a series of patches. The goal is to move
the fps_meter to autotest library and update the ebuild file so
that it is added to the /usr/local/bin and it can be used as a
tool accordingly. In this particular patch, fps_meter ebuild
files are removed.

BUG=chromium:845378
TEST=Manual tested

Change-Id: Ia0900ae6313875e527772a86d240065fea579148
Reviewed-on: https://chromium-review.googlesource.com/1074781
Commit-Ready: Yi Xu <yiyix@chromium.org>
Tested-by: Chung-yih Wang <cywang@chromium.org>
Reviewed-by: Hsu-Cheng Tsai <hctsai@chromium.org>