mesa: delay llvmpipe's raster threads creation
LLVMpipe's raster threads are created along with the display. This
causes a problem for Chromium because the threads are then created
before the GPU process has been sandboxed, which is a security
violation. Move the threads creation to the first render as a
workaround.

BUG=b:79891439
TEST=Checked that GPU process is working properly and that the raster
threads are also present.

Change-Id: Ifdb15339f565feccdd23a0c9514b12351244e241
Reviewed-on: https://chromium-review.googlesource.com/1078212
Commit-Ready: Alexandre Courbot <acourbot@chromium.org>
Tested-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>