git-eclass:  Simple check to validate git repository.
Recent failures have popped up related to a corruption of the git
repository for some git based packages.  This results in an inability to
build the impacted packages and relegates the entire tree dead.  This
will simply validate that an existing directory is a valid git repo,
removing the directory if not. The existing flow will handle the missing
directory and issue a fresh checkout.

BUG=chromium:846911
TEST=None

Change-Id: Id68d90866720c6bf0d0cb73dd931fa84134a490e
Reviewed-on: https://chromium-review.googlesource.com/1080069
Commit-Ready: Mike Nichols <mikenichols@chromium.org>
Tested-by: Mike Nichols <mikenichols@chromium.org>
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>