edk2: VRT patch
Subject: [PATCH] PcAtChipsetPkg: Maintain the RTC VRT bit

AMD SCO RTC VRT bit is updated every 256ms. The RTC logic
clears the bit when writing RegD. It then checks the bit to see if
the RTC is OK after a 10ms timeout and fails when it isn't set. We could
change the timeout, but that would delay the boot unnecessarily. Maintain
the VRT bit as it is updated regularly and carry on.

BUG=b:80230726
TEST=Boot Grunt to EFI shell.
BRANCH=none

Change-Id: Idb091e0fd68e17520a18d0abe26859749e830a59
Signed-off-by: Marc Jones <marc.jones@scarletltd.com>
Reviewed-on: https://chromium-review.googlesource.com/1089378
Commit-Ready: Martin Roth <martinroth@chromium.org>
Tested-by: Marc Jones <marc.jones@amd.corp-partner.google.com>
Reviewed-by: Martin Roth <martinroth@chromium.org>