cros-go.eclass: run tests with "-short" flag.
This is a standard flag supported by "go test". Typically, test files
check it to skip tests that take a long time, or tests that touch the
network. This should be the default behavior when running tests from an
ebuild. If an ebuild wants to run full tests, it can call "go_test"
directly with the desired package names (and without the -short flag).

BUG=chromium:852633
TEST=Emerged 56 ebuilds in chromiumos-overlay with FEATURES=test.
TEST='sudo FEATURES=test emerge dev-go/net' runs twice as fast.

Change-Id: I645f5c079aa577bdcbcf7ebd78d5d97275e2e366
Reviewed-on: https://chromium-review.googlesource.com/1101467
Commit-Ready: Rahul Chaudhry <rahulchaudhry@chromium.org>
Tested-by: Rahul Chaudhry <rahulchaudhry@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>