infineon-firmware-updater: Request TPM clear in recovery mode.
When handling a TPM firmware update retry in recovery mode, the
existing code that clears the TPM is known to fail because the TPM
won't handle the respective commands in bootloader mode. It's still
possible a TPM owner is present, and it'll survive if the retry
completes successfully. As a result, the SRK from before the update
would be carried over, which we must not allow. To address this,
request the firmware to preform a TPM clear unconditionally before
starting the update so we have reasonable confidence the TPM will be
in a good state after update installation and reboot.

BUG=chromium:854576
TEST=Manual, see bug.

Change-Id: Ief6e065ac7b2e5c987abfc335e13823c4a21702c
Reviewed-on: https://chromium-review.googlesource.com/1108117
Commit-Ready: Mattias Nissler <mnissler@chromium.org>
Tested-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>