cros-firmware: Support local firmware package without EC
At present with unibuild we require that the EC and PD iamges be present
in order to build a firmware image based on locally built binaries (as
opposed to 'official' ones downloaded from BCS).

Prior to unibuild the cros_ec USE flag was used to control this. Add this
same behaviour for unibuild.

This allows chromeos-firmware-scarlet to build. The problem was not
noticed for a long time since EC binaries were still present from before
it was converted to unibuild.

Note that we provide both EC and PD filenames to pack_firmware.py but it
checks the master config to see which ones are needed. For example, it
will only include the PD binary if the config has a 'pd-image' property.
If there is no 'pd-image' then the PD filename provided will be ignored,
and it is fine to provide is anyway.

BUG=chromium:854764
TEST=with go/croscl/1110025:
emerge-scarlet --unmerge chromeos-ec
FEATURES=test USE=-cros_ec emerge-scarlet chromeos-firmware-scarlet
See that this happily builds a local firmware image now and tests pass

Change-Id: Ideab9531c91b7c23f00d857d38ffdc09f870c2eb
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1110029
Reviewed-by: Nick Sanders <nsanders@chromium.org>