ps8751: add public firmware package for ps8751
In order to allow each board to control which version of firmware they
want, the entire package is masked by default. Each board can then
add a package.unmask with:

	=sys-firmware/parade-ps8751a3-55

This package will replace the package in private overlay as the firmware
is in the public bucket because we can freely distribute it. See
CL:*599246 for the addition of the private version.

BRANCH=none
BUG=b:78334391,b:79109245
TEST=verified firmware image is pulled into yorp FW image with full
stack of changes

Change-Id: I961a2a3243e4cffab103d2bd209356831efa17bd
Signed-off-by: Jett Rink <jettrink@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1110878
Reviewed-by: Raul E Rangel <rrangel@chromium.org>