Enable ThinLTO on all ARM boards.
We have enabled ThinLTO on elm and peach_pit in R68 and so
far so good. This CL enables ThinLTO on all ARM boards.

We enabled this feature via CL:1095415 but it got reverted because
of chromium:854801. We believe that we have identified the root
cause and fixed the problem.

BUG=chromium:827217 chromium:854801
TEST=cros tryjob daisy-release

Change-Id: I3a440f2faeff7c4ffa5a1117040f20ef21b63af0
Reviewed-on: https://chromium-review.googlesource.com/1120944
Reviewed-by: Yunlian Jiang <yunlian@chromium.org>
Commit-Queue: Yunlian Jiang <yunlian@chromium.org>
Tested-by: Yunlian Jiang <yunlian@chromium.org>