infineon-firmware-updater: Force TPM clear after update.
We need to make sure the TPM doesn't hold on to an SRK that originates
from before the TPM firmware update because that SRK might be weak.
This can happen in edge cases, for example when we start the TPM
firmware update using owner authorization but then get interrupted and
perform a retry via recovery mode.

This change adds post-update logic in the tpm-firmware-update.sh
script which runs during boot when a TPM firmware update has been
requested. It makes sure that we only continue booting if the TPM is
clear and requests the TPM to be cleared if necessary.

BUG=chromium:854576, chromium:778332, chromium:859511
TEST=See bug.

Change-Id: I091faf438e34c7c05944df1d26840367bc560fb7
Reviewed-on: https://chromium-review.googlesource.com/1123830
Commit-Ready: Mattias Nissler <mnissler@chromium.org>
Tested-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>