chromeos-base/ml: Ebuild changes to install models for ML Service.
We install both TFNative and TFLite models into rootfs.
This is the first way to use models in ML Service that we are
implementing. As this method does increase the storage in rootfs
it will be used to service-critical models only. For now we
want to test our first model (Tab Discarder).

For more info, see go/cros-ml-service-models.

BUG=chromium:836095
TEST=emerge chromeos-base/ml

Change-Id: Ie0f6c86cab61de656916b3ad41e94f3306b2bdea
Reviewed-on: https://chromium-review.googlesource.com/1125701
Commit-Ready: Claudio M <claudiomagni@chromium.org>
Tested-by: Claudio M <claudiomagni@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>