chromeos-cr50-scripts: start on started boot-services
Before this CL, a TPM2.0-specifc cr50-update upstart script waited for
login-prompt-visible, since it is a pre-condition for 'system-services'.
cryptohomed sits several links down the chain of daemons, which starts
after cr50-update. This cr50-update dependency led to the situation
when cryptohomed in TPM 2.0 case started much later than login prompt and
that caused delays with processing user passwords, if entered quickly.

This CL changes cr50-update dependency to 'started boot-services'.
That's the point where tpm-related daemons are started in TPM 1.2 case.
This changed allows reducing the delay from login-prompt-visible to
cryptohomed readiness for TPM 2.0 case to figures comparable with TPM
1.2 case (from ~4.5s to ~2.0/2.5s) while increasing time-to-login-prompt
by 0.05..0.15s only (1.6..3.2%), and still keeping time-to-login-prompt
shorter than in TPM 1.2 case. More details are provided in chromium:822302
and go/cryptohomed-startup-metrics.

This CL is a rare exception to the general rule, allowed due to the fact
that in terms of UX benefits this significant decrease of time to actual
login outweighs slight increase of time to login prompt.

BUG=b:67912719
BUG=chromium:822302
TEST=boot and login

Change-Id: I12344116d31c227f6b7d26d647a5e75f5c5e44b2
Reviewed-on: https://chromium-review.googlesource.com/724238
Commit-Ready: Andrey Pronin <apronin@chromium.org>
Tested-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>