sys-devel/llvm: Build llvm-config-host as a native build binary
Up until now, we've been building a version of LLVM for mesa that
both runs and produces binaries for the target architecture. One
wrinkle of that has been that mesa invokes llvm-config at compile
time to figure out how to compile with LLVM.

So far we've been using Qemu to emulate the ARM version of llvm-config.
This has started to cause problems, since we need to use Qemu's -L option
to point at /build/$BOARD and have llvm-config find its dynamic libraries
properly (probably just the interpreter). When there are symlink loops
in /build/$BOARD (as are created within the kernel build), Qemu gets lost
in an infinite directory traversal.

Emulating llvm-config is a suboptimal solution anyway. This change builds
llvm-config with all the same options as the ARM native compiler, but with
the build machine's CC and friends. This way we can stop with the Qemu
emulation.

BUG=chromium:807047
TEST=emerge-cheza ./llvm-6.0.0.ebuild; emerge-cheza mesa, observe no more hang
Also ran emerge-betty llvm mesa

Change-Id: I4cda717fa0d0fe4957454c31988250dcd4953b66
Reviewed-on: https://chromium-review.googlesource.com/1135511
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>