coreboot: Support the em100-mode USE flag on AMD boards
This USE flag reduces the speed of SPI flash so that em100 can be used to
hold the firmware. Adjust the ebuild to work with AMD as well. In this
case there is no ifdtool but we must pass a CONFIG option to the coreboot
build.

BUG=b:111363976
TEST=USE=em100-mode emerge-grunt -q  --nodeps coreboot chromeos-bootimage
See that it boots (em100 does not work yet)

Change-Id: I582f9a89e34bc25949e1f4ae70917718e3e5a325
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1135720