sys-apps/usbguard: Fix upstart config behavior, add whitelists.
The upstart config is now split into two config files since the old
config wouldn't behave properly if usbguard were to crash. With the
new config the whitelist is not regenerated if usbguard crashes,
and the default_authorized, and authorized flags are not enabled
when usbguard crashes but are enabled when the user logs back into
the device.

USE-gated whitelists were added for
 - detachable bases when hammerd is present.
 - keyboards for chromebox for meetings devices.

BUG=None
TEST=emerge and deploy usbguard and check the lockscreen USB behavior

Change-Id: I1e7d5f810d680a71e96a3f47ec042048d452be8f
Signed-off-by: Allen Webb <allenwebb@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1137060
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>