anx3429: add public firmware package for anx3429
In order to allow each board to control which version of firmware they
want, the entire package is masked by default. Each board can then
add a package.unmask with:

~sys-firmware/analogix-anx3429-21

This package will replace the package in private overlay as the firmware
is in the public bucket because we can freely distribute it.

BRANCH=none
BUG=b:79109245
TEST=install locally

Change-Id: I1f64d901f239de2a01c706efa801b2193cb3e3a0
Signed-off-by: Jett Rink <jettrink@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1144192
Reviewed-by: Raul E Rangel <rrangel@chromium.org>