libchrome: dbus: Add TryRegisterFallback
The TryRegisterFallback works just like TryRegisterObjectPath,
with addition that the registered callbacks also apply to the specified
object path and its sub paths. This is useful to implement a "catch-all"
handler.

Currently this is needed by Bluetooth dispatcher which needs to catch
all method calls to all objects and forward them to another D-Bus
service.

This is a cherry-pick of https://crrev.com/c/1134605.

BUG=chromium:862849
TEST=None

Change-Id: Iebbce5cb3a11b2fce601498043c1dd6e5d56790f
Reviewed-on: https://chromium-review.googlesource.com/1144190
Commit-Ready: Sonny Sasaka <sonnysasaka@chromium.org>
Tested-by: Sonny Sasaka <sonnysasaka@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>