usbguard: Use a stricter minijail0 policy.
Previously, seccomp was the only minijail feature used. This CL adds
the typical namespace flags as well as drops root and sets up a
mount namespace.

Minjail was added for "usbguard generate-policy", which is used to
generate the whitelist part of the rules.conf used by usbguard-daemon,
because it handles untrusted input (USB device descriptors).

CQ-DEPEND=CL:1169531
BUG=None
TEST=Verify /run/usbguard/rules.conf is generated on screen lock.

Change-Id: I98024571f4a63dac46ce21f2130a01acce94e5b3
Signed-off-by: Allen Webb <allenwebb@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1157296
Reviewed-by: Mike Frysinger <vapier@chromium.org>