cros-workon: drop cros-workon_src_install
This function doesn't do anything useful anymore: the callers using it
either do all the install they need to, or they already call `default`,
or they only call this function which is the same as `default`.  We have
never exported it by default either, so only people explicitly calling
it are impacted.  Drop it to simplify.

BUG=chromium:867926
TEST=precq passes

Change-Id: I7f200e83d20af3e083c497d3a0403c247678c6b0
Reviewed-on: https://chromium-review.googlesource.com/1163137
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>