cros-workon: drop cw_emake
This wrapper existed to help recover incremental build failures on bots.
The number of common.mk packages has declined significantly as we moved
to platform2/gyp, and our bots have been moving to GCE (so the hostname
check doesn't match) which means that logic doesn't kick in.  Lets drop
this to ease migration to a new common.mk eclass.

If we still run into incremental errors, we can add logic outside the
ebuild that nukes the caches for us.

BUG=chromium:867926
TEST=precq passes

Change-Id: Iafb71c04c37fc53be61104a9bdb19d8f949b3aed
Reviewed-on: https://chromium-review.googlesource.com/1163138
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Lann Martin <lannm@chromium.org>