target-chromium-os: drop old/redundant comments/embedded settings
The only reason for CROS_E_RDEPEND is to list util-linux, but we already
require that in virtual/implicit-system for its `mount` and related set
of tools.  Drop it here.

Also scrub old comments that apply to packages that have been moved or
dropped entirely.

BUG=None
TEST=precq passes

Change-Id: I74f38ef15bd4342b871e3ecaf0af00a2f582f969
Reviewed-on: https://chromium-review.googlesource.com/1164311
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>