edk2: Use fmap from first coreboot.rom file found
To get the size of the legacy partition, the ebuild finds the file
coreboot.rom and extracts the fmap information from it.  On
unibuild projects, there can be multiple rom files, so only look
at the first one and assume that all variants will use the same
size legacy partition.

BUG=b:72376807
TEST=emerge-grunt chromeos-bootimage builds edk2 into the image

Change-Id: I4256393cb2b981045c736346813cf3a4bfac0d9f
Signed-off-by: Martin Roth <martinroth@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1166227
Reviewed-by: Raul E Rangel <rrangel@chromium.org>