Adding patch to allow PWG Raster types to vary by case
This change adds a patch to CUPS which fixes an issue where Epson
printers would always print in black and white when provided with
pwg-raster, even if the requested print was in colour.

Their reported color scheme was not recognized by CUPS, causing
them to default to black and white.

BUG=chromium:870099, chromium:804990
TEST=Tested Manually on device

Change-Id: I67473fa41a3a201e6e59312ec17669d8e34bb35a
Reviewed-on: https://chromium-review.googlesource.com/1170082
Commit-Ready: David Valleau <valleau@chromium.org>
Tested-by: David Valleau <valleau@chromium.org>
Reviewed-by: Sean Kau <skau@chromium.org>