libchrome: dbus: Remove LOG(ERROR) in ObjectProxy
It is a valid use case for a daemon to have multiple ObjectProxies of
different services with the exact same object path and interface name.
Currently, this may cause log pollution of "rejecting a message from a
wrong sender" because one ObjectProxy receives signals intended for
another ObjectProxy. Since it's actually a valid case and not a bug, it
shouldn't be logged as error but it may still be logged with VLOG.

Currently this is discovered in Bluetooth daemon (btdispatch) because it
listens to both BlueZ's and Newblue's objects which have identical
object paths and interfaces.

This is a cherry-pick of https://crrev.com/c/1164474.

BUG=chromium:866704
TEST=Build libchrome, test with btdispatch and check that there is no
more log pollution

Change-Id: Ib4224d3c639e87adb7897e02956215fee513a879
Reviewed-on: https://chromium-review.googlesource.com/1170190
Commit-Ready: Sonny Sasaka <sonnysasaka@chromium.org>
Tested-by: Sonny Sasaka <sonnysasaka@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>