Explicitly disable X11.
It is no longer necessary to maintain project-freon as a separate
entity from the main overlay, so it is being merged into the defaults.
This commit re-adds the freon setting that explicitly disables X11.
The disable was previously removed along with the USE enabling X11,
but that is not sufficient, it needs to be explicitly disabled.

BUG=chromium:730026
TEST=local builds, precq passes

Change-Id: Ice300a3e3896548d98daf917022d207537b3ab31
Reviewed-on: https://chromium-review.googlesource.com/1173411
Commit-Ready: Alex Klein <saklein@chromium.org>
Tested-by: Alex Klein <saklein@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>