usbguard: Rename soraka whitelist to hammer and add whiskers.
This tightens the whitelisting rules for the detachable bases
to include the vendor id and device name. New rules were added
for whiskers, and for the RO firmware.

This also adds copyright notices to all the rules files.

"sudo usbguard generate-policy" lists a set of rules for
connected devices. This can be used to help write rules
for whitelists such as the ones for hammer derivatives.

BUG=b:112621585
TEST=Connect whiskers while at lockscreen and verify it works.

Change-Id: Ia800afa47ed770f9491cd677f830e8f3c92863a3
Signed-off-by: Allen Webb <allenwebb@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1176285
Tested-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>