libdrm, arc-libdrm: Switch to using epatch and a single branch
libdrm and arc-libdrm used a custom packaging scheme where we had a
branch off of the upstream tag with the downstream patches. For
example, chromeos-2.4.92 was branched off of the upstream
libdrm-2.4.92 tag and had a few commits on top with our downstream
additions.  Rebasing to a new libdrm was awkward and involved creating
a new chromeos libdrm branch, cherry-picking downstream patches
through the CQ, updating the manifests.

This CL switches the ebuild over to use epatch and a regular branch,
much like how we package the mesa project. This makes rebasing much
simpler, with the downside of making adding a downstream patch
slightly more work. If anything, we should favor upstream work and
rebasing over local patches, so this is a reasonable tradeoff.

BUG=b/112720782
TEST=repo sync, emerge libdrm and arc-libdrm
CQ-DEPEND=CL:*667008,CL:1188443

Change-Id: I65fad7a46d9350ba3603fc9801f5ab6765edc06d
Reviewed-on: https://chromium-review.googlesource.com/1188684
Tested-by: Kristian H. Kristensen <hoegsberg@chromium.org>
Reviewed-by: Kristian H. Kristensen <hoegsberg@chromium.org>