libchrome: Cherry-pick r396638.
Time::From*Exploded() is used in ChromeOS and the API is changed
in the next uprev. With this Cherry-pick, clients can be updated
in advance to the coming uprev.

Note: to avoid patch failure, time_win.cc is manually removed
just because the file does not exist in libchrome.

BUG=b:37434548
TEST=Ran trybot.

Change-Id: Id202dc1aae0b24bc139c0a08e2a8de70582b152b
Reviewed-on: https://chromium-review.googlesource.com/1195205
Commit-Ready: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>