chromeos-ec: Install npcx_monitor.bin into firmware directory
npcx_monitor.bin is required when flashing the NCPX via a Suzy-Q cable.

This is added to the chromeos-ec build because the .bin will only exist
for npcx chips and should be included in the firmware tarball so that
the updateuart tool can use it.

It was not added as part of ec-devutils because that package is board
agnostic.

BUG=b:113288489
TEST=emerge-grunt chromeos-ec and validated the file was installed

Change-Id: I7dad341232051348e977ee775186084678b411d4
Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1194646
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Jett Rink <jettrink@chromium.org>