lld: get the real binary in the wrapper.
We created a wrapper for lld so that we can add parameters for lld
more easily. To do that, we moved the binary 'lld' to lld.real and created a
wrapper 'lld' to call lld.real
In the wrapper, we use
  exec ${0}".real to call the real bianry
There is a symbolic link (ld.lld) points to lld. If some one uses ld.lld,
it will look for ld.lld.real, which does not exist, an error will be poped out.
To fix this issue, we use
  exec "$(readlink -f "${0}").real" to call the real binary.

BUG=None
TEST=x86_64-cros-linux-gnu-clang++ -fuse-ld=lld works.

Change-Id: I41513795f2d3f9112237abd0517c22f3da3ab1b0
Reviewed-on: https://chromium-review.googlesource.com/1194649
Commit-Ready: Yunlian Jiang <yunlian@chromium.org>
Tested-by: Yunlian Jiang <yunlian@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Yunlian Jiang <yunlian@chromium.org>