sys-kernel/linux-headers: Move V4L2_CTRL_FLAG_CAN_STORE far from previous flag
The V4L2_CTRL_FLAG_CAN_STORE flag was added by a downstream CL:232581,
which assigned it a bit next to the previous flag. It was then moved one
bit further due to a collision with upstream, which is now the case
again. Move if far a way from previous flag, so the chance of this
happening again is much lower.

Note that even though this flag is exposed to user space, it is not used
currently by our user space code, so there is no compatibility concern.

This is a header change corresponding to following kernel CL:1194553 and
CL:1195217.

BUG=chromium:878124
TEST=pre-CQ passes (unused definition)

Change-Id: Ic359877506308e79e21cbc0417fae42995633f6c
Reviewed-on: https://chromium-review.googlesource.com/1194530
Commit-Ready: Tomasz Figa <tfiga@chromium.org>
Tested-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-by: Pawel Osciak <posciak@chromium.org>